Update README.md
# python-exercises-solved
Collection of solved exercises of programming with python most of them are exercises applied to economics, but there are also generic exercises for any discipline.

# What will I learn?
This experience is integrated into the Level 1 Problem Solving employability competency, where a problem will be identified and analyzed to generate solution alternatives, applying the methods learned. For this, the student must:

Recognize what a problem is, explaining it before tackling it.
Ask questions to define the problem posed, limiting its scope and impact.
Apply a logical method in solving the problem, considering defined and interrelated steps.
Apply a logical method in solving the problem, considering defined and interrelated steps.

To begin the second learning experience, the first steps for programming in the Python Language will be announced, teaching the code and basic concepts of programming to students, printing values ​​on the screen, storing variables, requesting variables on the screen, among others, a Once introduced in the language, it will be reinforced with exercises of the proposed activity where an order is required when programming and using tabulation tools and line breaks for a better code presentation, then we will see the simple conditions and nested conditions, where students must apply all their logical knowledge for their understanding, using truth tables (basic level) where we will use unions and intersections of conditions, for comparison. We will analyze the different cases and of course, we will see the great importance of indentation in Python. Once the above has been analyzed, it will be reinforced with a training activity, in order for students to manage this resource, since it will be one of the most used during their programming subjects. To conclude, with this learning experience, variables of the counter type, flags and accumulators will be seen, explaining to the students the difference between them and the importance of each one, improving logical thinking to develop good programming practices.
# MANDATORY BIBLIOGRAPHY

Python Getting Started, (1999-2020). Rescatado de https://www.w3schools.com/Python/python_getstarted.asp 
Python Comments, (1999-2020). Rescatado de https://www.w3schools.com/Python/python_comments.asp 
Python Conditions and If statements, (1999-2020). Rescatado de https://www.w3schools.com/Python/python_conditions.asp